either-kotlin: rename .a and .b to .leftValue and .rightValue
Having too generic and vague .a and .b fields in Kotlin class is bad,
because with current approach and data classes, the fields are
accessible to users of the library.  As such, the `when` construct for
pattern matching looks a bit confusing when `e.a` and `e.b` are used.

So rename them to .leftValue and .rightValue for now.