IDEA: add kotlinc.xml
Similar to JDK version in .idea/misc.xml, it seems correct to keep
configuration file of Kotlin Plugin/Compiler -- kotlinc.xml -- in
repository.  As of 2023, IntelliJ IDEA itself keeps it in Git.[1]

So commit whatever version of this config file that I have locally right
now.  The version string (of Kotlin Plugin?) in it seems to correspond
to whatever I've put into either-kotlin/build.gradle.kts, and IntelliJ
IDEA (version 2022.3.1) doesn't seem to mind it.

[1] https://github.com/JetBrains/intellij-community/blob/30fa6f95fa155f88ffe1f6ce93eebd3ea7ce9162/.idea/kotlinc.xml