Gradle: set up publishing
To be able to use libraries of Either-JVM, they need to be able to be
published.  Set up basic configuration for the Gradle plugin
`maven-publish`.  Note that instead of using operator []:

    https://github.com/gradle/gradle/blob/330b2d72f5e9f692de92697cb580e93fc6e82545/subprojects/kotlin-dsl/src/main/kotlin/org/gradle/kotlin/dsl/NamedDomainObjectCollectionExtensions.kt#L391-L401

as the sample code for publishing recommends:

    https://docs.gradle.org/7.6/userguide/publishing_setup.html#sec:basic_publishing

I use `getByName` directly, which makes it more obvious what's going on.
At the moment IntelliJ IDEA doesn't allow navigation to the declaration
of the extension operator function `get`.