either-java*: either(f,g,e): allow subclasses as types in Either parameter
If user of method `Either.either` that takes three parameters forces the
type parameters, we should be flexible in the third parameter of type
`Either`.  Replace plain type parameters `<A, B>` with bounded wildcard
type parameters `? extends A, ? extends B`.

Here, for the purposes of the PECS mnemonic [1], the producer is the
third parameter `e`.  An `Either<? extends A, ? extends B>` produces
elements of type `A` or `B`.

[1] https://stackoverflow.com/a/2723538/1083697