either-java*: match(): allow superclasses as input types in parameters
Users of APIs of class Either might want to pass method references as
parameters of type `Function`.  This might cause problems in cases,
where a method takes a superclass of the values used in the type
parameters of Either.  See new tests for concrete examples of how this
might happen: using Number instead of Integer and using CharSequence
instead of String.

Follow the "Consumer---Super" side of the PECS mnemonic [1] and replace
plain type parameters `A` and `B` with `? super A` and `? super B`
correspondingly in the parameters of method `match()`.

Here, for the purposes of the PECS mnemonic, the consumers are the
functions `f` and `g` -- parameters of method `match()`.  They consume
inputs of type `A`.

[1] https://stackoverflow.com/a/2723538/1083697