either-java*: match(): allow subclasses as output types in parameters
Users of APIs of class Either might want to pass method references as
parameters of type `Function`.  This might cause problems in cases,
where a method returns a subclass of `R` which is the type parameter of
method Either#match.  See new tests for concrete examples of how this
might happen: using `List<String>` instead of `Collection<String>`.

Follow the "Producer---Extends" side of the PECS mnemonic [1] and
replace plain type parameters `R` with `? extends R` in the parameters
of method `match()`.

Here, for the purposes of the PECS mnemonic, the producers are the
functions `f` and `g` -- parameters of method `match()`.  They produce
output of type `R`.

[1] https://stackoverflow.com/a/2723538/1083697