either-java*: either(): allow subclasses as output types in Function parameters
Users of APIs of class Either might want to pass method references as
parameters of type `Function`.  This might cause problems in cases,
where a method returns a subclass of `C` which is the type parameter of
method Either#either.  See new tests for concrete examples of how this
might happen: using `List<String>` instead of `Collection<String>`.

Follow the "Producer---Extends" side of the PECS mnemonic [1] and
replace plain type parameters `C` with `? extends C` in the parameters
of method `either()`.

Here, for the purposes of the PECS mnemonic, the producers are the
functions `f` and `g` -- parameters of method `either()`.  They produce
output of type `C`.

[1] https://stackoverflow.com/a/2723538/1083697