*: drop type classes for now
I was too ambitious when I added methods for Functor, Applicative, etc
to the libraries.  At the moment I'm trying to polish all the other
methods and functions, Gradle build scripts, and documentation in
preparation of publishing the libraries.  All of these type class
methods are just distracting and confusing at the moment.

Drop the implementations of type class methods for now.

They might be added back in the future, most likely as separate
libraries, something like `either-java17-type-classes`.