either-kotlin: Either: add method `match`
Using builtin `when` expressions in Kotlin [1] might be convenient for
some use cases, but it is cumbersome in others.  Sometimes the user
might want to use a variable of function type or a function/method
reference [2] in the branches of the pattern matching.  Using functional
variables or method references in `when` expression requires spelling
out the invocations of those functions or methods in the branches of the
`when` expression, aka the pointful style, the opposite of pointfree.[3]

Using the `when` expression also requires a bit of boilerplate:

  - the branches `is Either.Left` and `is Either.Right`
  - the access of the fields `.leftValue` and `.rightValue`

which reduces readability.

Introduce method `match()` in class Either of either-kotlin, which takes
two functions.  This is similar to the methods `match()` in Java
variants of class Either.

[1] https://kotlinlang.org/docs/control-flow.html#when-expression
[2] https://kotlinlang.org/docs/reflection.html#function-references
[3] https://wiki.haskell.org/Pointfree