*: add accept() and peek()
Method match() requires the functions passed in as parameters to return
a value.  But in imperative languages like Java and Kotlin, returning a
value is not always necessary.

Add methods accept() and peek() that take two `Consumer`s (functions
that return `Unit` in case of Kotlin) to correspond to the existing
method match().

Naming method peek() was very easy.  The name was just lifted straight
from java.util.stream.Stream#peek, which does the operation similar in
spirit to what Either#peek does.

On the other hand, I'm not sure about the name "accept" -- it is just
taken from `java.util.function.Consumer#accept`, but it is actually an
operation that is "dual" to `Consumer#accept`.  Classes like Optional
and Iterable/Collection have the benefit of using context for naming:

  - java.util.Optional#ifPresent
  - java.lang.Iterable#forEach

but there isn't a good way to refer to what accept() does for Either.
The word "apply" is already taken by `java.util.function.Function#apply`.