either-kotlin: mark either() as `inline`
Similarly to previous commit for `match()`, mark two variants of the
`either()` function as `inline`.  For explanation of `crossinline`
modifier in the second variant, see
https://stackoverflow.com/a/65893906/1083697
This is needed, because `f` and `g` are called not inside `either()`
butr inside a separate execution context, i.e. the lambda that is
returned from `either()`, that may be executed somewhere outside of the
control of the code of `either-kotlin`.