java17: use sealed interface and records
It ism't possible to use records for `Left` and `Right` with a sealed
class, because records cannot extend other classes.  But class `Either`
doesn't have any state.  This means that we don't really need a class.
An interface is enough.  And we can also make the interface sealed!

Replace usage of sealed class in either-java17 with a sealed interface
for class Either.  Make subclasses Left and Right into records.  It also
removes the need to define methods equals() and hashCode() separately.