java8 & kotlin: use separate packages
We have to separate classes of either-java17 into a separate package to
avoid all kinds of problems.  It allows using either-java8 and
either-java17 in one project.  Having classes with the same name in the
same package also confuses IntelliJ IDEA, as mentioned previously in
commit 5581e43 (either-java17: move classes to separate package,
2021-08-08).

So move classes of either-java8 and either-kotlin into corresponding
packages.  This would allow using either-java8, either-java17, and
either-kotlin in the same project.