Gradle: checkSpdxLicenseIdentifiers: new task
As recent commits have shown, I often forget to add the comment with
SPDX-License-Identifier to source code files.

Add an automatic check for presence of SPDX-License-Identifier comments
as a new Gradle task checkSpdxLicenseIdentifiers.  For now, it only
checks Java and Kotlin source files.

Use a dummy file 'build/reports/license-check-marker.txt' to make the
Gradle's UP-TO-DATE detection work.  I'm not sure if this is a good way
to do it, but it's good enough for now.