*: improve mentions of type parameters in Javadocs and KDocs
Javadoc comments in either-java* format mentions of type parameters
using `{@code}`.  Correspondingly, KDoc comments format mentions of type
parameters using `monospace`.

Use `{@link}` and [link] syntax instead to highlight the types involved
better and to allow users to click on the rendered link to see the
description of the type parameters.[1]  Add missing parameter
descriptions for static methods and companion functions.[2]

While we're here, update mentions of instances of type `Either<A, B>` in
the surrounding text and tweak wording for documentation comments of
static methods (Java) and functions (Kotlin) left() and right().

[1] There is a bug in IntelliJ IDEA 2023.1, where clicking only works
    for KDoc and not for Javadoc:
    https://youtrack.jetbrains.com/issue/IDEA-326131/Quick-documentation-of-type-parameters-in-Java-doesnt-lead-to-their-description
[2] Note that static methods don't share type parameters with the class
    definition.