*: fix mentions of type parameters in comments
Javadoc doesn't support linking to type parameters in comments.  It
seems to be a bug in Javadoc that it doesn't produce an error on such
input.  Previous commit [1] was incorrect in assuming that this
was possible.

For reference, documentation of Javadoc [2] has the following:

    {@link reference label}

    [...] points to the documentation for the specified module, package,
    class, or member name of a referenced class.

So replace all {@link}s for type parameters with {@code}.  Similar issue
exist in Dokka.  The links generated for type parameters by Dokka do
have the correct text, but their URL just leads to the current page.
So fix the same issue in KDocs as well.

[1] 9507b80 (*: improve mentions of type parameters in Javadocs and
    KDocs, 2023-07-24)
[2] https://docs.oracle.com/en/java/javase/20/docs/specs/javadoc/doc-comment-spec.html