kotlin: apply Gradle plugin for Dokka
Add Gradle plugin for Dokka, the documentation engine for Kotlin, so
that we can later generate documentation for publishing.  At first, it
will be just in javadoc style (which Dokka can also generate), but
hopefully I'll get a hang of Kotlin's ways of doing things and figure
out how to do non-javadoc-style documentation for either-kotlin.

- https://kotlinlang.org/docs/dokka-gradle.html
- https://kotlinlang.org/docs/dokka-introduction.html