java*: EitherStreams: tweak @return in javadocs
The way "{@link Stream} of {@code A}" is rendered in IntelliJ IDEA,
makes "Stream" a blue link and "of A" is white (on a dark grey
background).  This makes it a bit hard to parse.

Replace "Stream of A" with "Stream<A>" (and the same for type "B") in
@return of javadocs of class EitherStreams.  Notation "Stream<T>" should
be clear to most Java developers.