Gradle: merge configuration of publishing
Merge configuration of publishing in subprojects into an ad-hoc plugin
`either-jvm-build.maven-publish-conventions`.  Because the maven
publications are split between subprojects, we don't need the specific
names for them.  The new plugin also assumes that Java and Kotlin code
aren't mixed in any of the users of the plugin.

See also JUnit 5's publishing conventions:
    https://github.com/junit-team/junit5/blob/main/gradle/plugins/common/src/main/kotlin/junitbuild.publishing-conventions.gradle.kts